\name{LRTT-package}
\alias{LRTT-package}
\alias{LRTT}
\docType{package}
\title{
\packageTitle{LRTT}
}
\description{
\packageDescription{LRTT}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{LRTT}
\packageIndices{LRTT}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{LRTT}

Maintainer: \packageMaintainer{LRTT}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
