% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tree_ratio.R
\name{Tree.ratio.back}
\alias{Tree.ratio.back}
\title{correct step for tree ratio test}
\usage{
Tree.ratio.back(tree.ratio)
}
\arguments{
  \item{tree.ratio}{results return by Tree.ratio function}
}
\value{
   The differential OTU detected
}
\description{
Correct step for tree ratio test
}
\seealso{
\code{\link{Tree.ratio}}
}
