% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tree_ratio.R
\name{Tree.ratio.back}
\alias{Tree.ratio.back}
\title{correct step for tree ratio test}
\usage{
Tree.ratio.back(p, tree.ratio, taxa.index, otutab, group = c(1:N))
}
\arguments{
  \item{p}{OTU number}
  \item{tree.ratio}{results return by Tree.ratio function}
  \item{taxa.index}{results return by Taxa.index function}
  \item{otutab}{OTU table}
  \item{group}{group label}
}
\value{
   The differential OTU detected
}
\description{
Correct step for tree ratio test
}
\seealso{
\code{\link{Tree.ratio}}
}
